feat: finalizados os testes no módulo de administradores. Todos os testes estão funcionando corretamente, sem erro nenhum. Adicionada uma validação no método de excluir no AdministradorService para verificar se o id existe antes de excluir e lançar exceção se for o caso

+++ src/main/java/com/project/ecopraia/service/AdministradorService.java
@@ -65,6 +65,10 @@ public class AdministradorService {
     }
 
     public void excluir(Long id) {
+        if(!administradorRepository.existsById(id)) {
+            throw  new RuntimeException("Administrador não encontrado");
+        }
+        
         administradorRepository.deleteById(id);
     }
 }
